update do user feito

+++ Project/database/DeleteUPP.php
@@ -12,16 +12,19 @@ $stmt->execute(array($val));
 $deleteId = $stmt->fetch();
 $deleteId = $deleteId['photoId'];
 
-$stmt = $db->prepare("UPDATE User SET photoId = NULL WHERE username = ?");
-$stmt->execute(array($val));
+if($deleteId != null)
+{
+	$stmt = $db->prepare("UPDATE User SET photoId = NULL WHERE username = ?");
+	$stmt->execute(array($val));
 
-$stmt = $db->prepare("DELETE FROM Photo WHERE photoId = ?");
-$stmt->execute(array($deleteId));
+	$stmt = $db->prepare("DELETE FROM Photo WHERE photoId = ?");
+	$stmt->execute(array($deleteId));
 
-unlink("../css/images/$deleteId.jpg");
-unlink("../css/images_small/$deleteId.jpg");
-unlink("../css/images_medium/$deleteId.jpg");
+	unlink("../css/images/$deleteId.jpg");
+	unlink("../css/images_small/$deleteId.jpg");
+	unlink("../css/images_medium/$deleteId.jpg");
+}
 
-header('Location: ../restaurantProfileEdit.php?restaurant=' . $val);
+header('Location: ../userProfileEdit.php?user=' . $val);
 exit;
 ?>
